refactor: fold MCP App support into directions_tool (drops directions_app_tool)
Replaces the standalone directions_app_tool with a single source of truth:
directions_tool now serves the live GL JS map for both MCP Apps (via
_meta.ui.resourceUri pointing to DirectionsAppUIResource) and legacy MCP-UI
(via inline rawHtml UIResource on the tool result, gated by ENABLE_MCP_UI).

- Extract HTML template into directionsAppHtml.ts shared by both pathways
- MCP Apps: resource serves generic HTML, iframe receives geometry via
  postMessage from ui/notifications/tool-result
- MCP-UI: tool inlines the same HTML with geometry baked in via an
  initial-data <script> block, so non-MCP-App clients also render GL JS
- Non-supporting hosts ignore both UI hints and consume the existing
  text/structuredContent payload unchanged
- Public token resolved server-side via Tokens API or MAPBOX_PUBLIC_TOKEN

No more "GL JS for one client, static image for the other."
